Create Chat-View
This is the MVC View for the chat. It takes in a list of Chat Message objects queried from a database, and populates a section of the page using a foreach loop. New messages are sent as an action attached to a button and sent to a server, which in-turn calls the addMessage() function on this View.

Javascript did not have any readily available functions to easily format the date/time the new messages the same. This was solved by rendering the date/time as a string on the server and passed back as a simple javascript string.